Add DepthImageProcessor class and unit test.
Summary: Since EFM needs Depth images, this diff adds DepthImageProcessor class. In the mean time, Depth images will need to perform the same transformations as the raw images, except that the interpolation method needs to be NEAREST. Hence I exposed the transformation as a list of `torchvision.transforms` out of `AriaCameraProcess` class to reuse across processors.

Reviewed By: hlzz

Differential Revision: D57146859

fbshipit-source-id: d90eca438d79cd71cd9e9a510a00db6e15260815