test WDS writing for MpsOnlineCalibData
Summary:
## Context

The stack of diffs is focused on adding new features and functionality to the MpsDataProvider library, specifically related to online calibration data. The overall purpose of the stack is to enhance the capabilities of the MpsDataProvider library for use in various applications.

## This Diff

This diff is titled "[ATEK] test WDS writing for MpsOnlineCalibData" and contains code changes in the file "atek_data_sample_test.py". The changes include adding a new class "MpsOnlineCalibData" to the file and testing the writing of this class to a Web Data Service (WDS) using the "write_to_wds" function. The changes are intended to improve the functionality of the ATEK data preprocessing pipeline.

Reviewed By: YLouWashU

Differential Revision:
D59832249

Privacy Context Container: L1257120

fbshipit-source-id: 498781d786b9f663ba608c4f4b8d9cde816f6e02