WDS writing for MpsOnlineCalibData
Summary:
## Context

The stack of diffs is focused on adding new features and functionality to the MpsDataProvider library, specifically related to online calibration data. The overall purpose of the stack is to enhance the capabilities of the MpsDataProvider library for use in various applications.

## This Diff

This diff introduces a new function `to_flatten_dict()` to the `atek_data_sample.py` file. This function flattens the fields of a given object into a dictionary, skipping any fields that are None or empty strings. This function will be useful for converting the `MpsOnlineCalibData` object into a flattened dictionary, which can then be written to a file using the `WDS` writing function.

Reviewed By: YLouWashU

Differential Revision:
D59832251

Privacy Context Container: L1257120

fbshipit-source-id: d3a99ebd4007ffc414a4d0939ac03acf27f4ce51