Add a flag for labeled data in `get_image_transform` to support reading segmentation data
Summary:
When trying to apply the camera transform to ADT's segmentation masks, the undistortion to a linear camera would alter values, making the mask invalid and impossible to pull out each segmentation.

This adds an additional parameter that lets us read segmentation info correctly from DepthImageProcessor.

Note that now DepthImageProcessor is completely compatible with segmentation, and there wasn't a need to make a SegmentationImageProcessor; maybe it should be renamed to have a more general name?

Reviewed By: YLouWashU

Differential Revision: D60528055

fbshipit-source-id: f69b9b794f2a2c3fbb9cfd096edbc6d2d670362a