fix bug: multithreading error in visualization for inference result.
Summary:
For inference, we load data to multiple CPUs for multithreading. However, the visualization should be single-threaded. For example, if we set the number of workers to 2, the trajectory will be separated into two lines, and every worker will update the whole trajectory by adding points one after another, causing a zig-zag effect. Now I have added a flag so that if the user wants to turn on visualization, the inference will become single-threaded.

Before: {F1804418641}
After:  {F1804427094}

Reviewed By: YLouWashU

Differential Revision:
D61245944

Privacy Context Container: L1257120

fbshipit-source-id: d26855d11ecd7ccdb4d489a77b9f7688cad7b894